pyproject.toml
[project]
name = "shiva-ai"
version = "1.0.0"
description = "Desi AI Bhai - A beautiful chatbot powered by Google Gemini"
requires-python = ">=3.9"
authors = [{name = "Shiva", email = "shiva@example.com"}]
license = {text = "MIT"}

dependencies = [
    "flask==3.0.0",
    "flask-cors==4.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
